make processing deferred routes rd aware
In attempting to enable graecful restart for EVPN routes, I found that
when routes were deferred waiting for an end-of-rib marker from all of
the GR-enabled peers, they were never processed. I found that this was
because the code to process deferred routes was not aware of route
distingusihers in the RIB and how they refer to a second level of
individual routing tables.