ensure that the node gets unlocked in all cases
The logic to unlock dest if iteration completed without iterating the
entire node was a bit flawed. For example, if iteration terminated due
to `gr_deferred == 0` then the node would not get unlocked.

This change takes into account the fact that dest will be NULL only in
the case when the entire table was iterated and all nodes were already
unlocked. In any other case, it needs to be unlocked.