separate sigint and sigterm and modify sigint to skip BGP shutdown
Unlike gobgp and bird – neither of which send a BGP notification to
peers when going down – FRR uses the same signal handler for both
SIGINT and SIGTERM. This handler ends up sending two notifications:

- Peer de-configurated
- Administratively shutdown

These notifications interfere with the operation of graceful restart
on the peer-side (acting as GR helpers). I’m not sure what the
intended way to restart FRR is to trigger GR in helper mode on its
peers.

This change makes FRR skip BGP shutdown when a SIGINT is received.
This might require a bit more finess in the long-term but for now, it
seems to trigger graceful restart in peers.